feat(models): move benchmark model generator code
