+++ lib/codegen/codegen.js
@@ -36,7 +36,7 @@ const JSONSchemaToConcertoVisitor = require(
 const OpenApiToConcertoVisitor = require('./fromOpenApi/cto/openApiVisitor');
 const RustVisitor = require('./fromcto/rust/rustvisitor');
 const VocabularyVisitor = require('./fromcto/vocabulary/vocabularyvisitor');
-const BenchmarkModelGenerator = require('./benchmarkModelGenerator/benchmarkModelGenerator');
+const BenchmarkModelGenerator = require('../common/benchmarkModelGenerator');
 
 module.exports = {
     AbstractPlugin,
@@ -76,6 +76,5 @@ module.exports = {
         avro: AvroVisitor,
         rust: RustVisitor,
         vocabulary: VocabularyVisitor,
-        benchmarkModelGenerator: BenchmarkModelGenerator,
     }
 };
